pyproject.toml
[project]
name = "langchain-multi-vector-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.13",
    "langchain-openai>=0.2.14",
    "langchain-postgres>=0.0.12",
    "langchain>=0.3.13",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
]
